それっぽくなってきた

+++ src/ick/base/crt.cpp
@@ -66,8 +66,8 @@ namespace ick{
 		va_copy(ap2, ap);
 		int num = ick::vsnprintf(NULL, 0, format, ap2);
 		va_end(ap2);
-		size_t size = static_cast<size_t>(num + 1);
-		*ret = ICK_ALLOC_A(allocator, char, size);
+		size_t size = static_cast<size_t>(num+1);
+		*ret = ICK_ALLOC_A(allocator, char, static_cast<int>(size));
 		return ick::vsnprintf(*ret, size, format, ap);
 	}
 }
